# refactor random-ipv4addrs.py for faster execution